enhance password errors
When fixing return errors from the server I noticed that I could give a better message if the name/password were not authorized. So now I post a better error in the logs for this common case.

I foound these erros when ryobi was timing out the other day when getting device info. I hope that was an issue on there end or some and not specfically blocking my request. There whole backend was not down at the time so it is hard to knkow for sure.

If however they start to actually block me from pollig then....i'd have to put this project to rest. :-(